# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/posts": {
      "get": {
        "operationId": "PostController_getPosts",
        "parameters": [
          {
            "name": "tags",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull."
          }
        },
        "summary": "Get posts.",
        "tags": [
          "Post"
        ]
      },
      "post": {
        "operationId": "PostController_createPost",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "401": {
            "description": "Unnauthorized."
          },
          "403": {
            "description": "Forbidden."
          }
        },
        "summary": "Create post.",
        "tags": [
          "Post"
        ]
      }
    },
    "/posts/{id}": {
      "get": {
        "operationId": "PostController_getPostById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Post not found."
          }
        },
        "summary": "Get post with id or slug.",
        "tags": [
          "Post"
        ]
      },
      "patch": {
        "operationId": "PostController_updatePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Post not found."
          }
        },
        "summary": "Update post.",
        "tags": [
          "Post"
        ]
      },
      "delete": {
        "operationId": "PostController_deletePost",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successfull."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Post not found."
          }
        },
        "summary": "Update post.",
        "tags": [
          "Post"
        ]
      }
    },
    "/posts/{id}/slug": {
      "patch": {
        "operationId": "PostController_changeSlug",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeSlugDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Post not found."
          },
          "409": {
            "description": "Conflict."
          }
        },
        "summary": "Update slug.",
        "tags": [
          "Post"
        ]
      }
    },
    "/posts/status/{id}": {
      "patch": {
        "operationId": "PostController_changeStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeStatusPostDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Post not found."
          }
        },
        "summary": "Update status.",
        "tags": [
          "Post"
        ]
      }
    },
    "/posts/{id}/comments/count": {
      "get": {
        "operationId": "PostController_GetCommentCount",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull."
          }
        },
        "summary": "Get comment count.",
        "tags": [
          "Post"
        ]
      }
    },
    "/posts/{postId}/comments": {
      "post": {
        "operationId": "PostController_AddComment",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfull."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Post not found."
          }
        },
        "summary": "Create a comment.",
        "tags": [
          "Post"
        ]
      },
      "get": {
        "operationId": "PostController_getComments",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfull."
          },
          "404": {
            "description": "Post not found."
          }
        },
        "summary": "Get Comments.",
        "tags": [
          "Post"
        ]
      }
    },
    "/posts/{postId}/tags/{tagId}": {
      "post": {
        "operationId": "PostController_AddTag",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successfull."
          },
          "400": {
            "description": "Bad Request."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden."
          },
          "409": {
            "description": "Conflict."
          }
        },
        "summary": "Create tag.",
        "tags": [
          "Post"
        ]
      },
      "delete": {
        "operationId": "PostController_RemoveTag",
        "parameters": [
          {
            "name": "postId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfull."
          },
          "403": {
            "description": "Forbidden."
          },
          "404": {
            "description": "Post not found."
          }
        },
        "summary": "Delete tag.",
        "tags": [
          "Post"
        ]
      }
    },
...
  },
  "info": {
    "title": "Blog API",
    "description": "API Documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ChangeSlugDto": {
        "type": "object",
        "properties": {}
      },
      "ChangeStatusPostDto": {
        "type": "object",
        "properties": {}
      },
      "CreateCommentDto": {
        "type": "object",
        "properties": {}
      },
      "UpdatePostDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}